moved Ui.java back into dolla package (idk how to resolve the conflict)
