+++ src/main/java/dolla/Storage.java
@@ -6,7 +6,6 @@ import dolla.parser.MainParser;
 import dolla.task.Debt;
 import dolla.task.Entry;
 import dolla.task.Log;
-import dolla.ui.Ui;
 
 import java.io.BufferedReader;
 import java.io.FileNotFoundException;
