+++ src/main/java/dolla/command/AddDoAfterTaskCommand.java
@@ -1,6 +1,6 @@
 package dolla.command;
 
-import dolla.ui.Ui;
+import dolla.Ui;
 import dolla.task.DoAfter;
 import dolla.task.Task;
 import dolla.task.TaskList;
