+++ src/main/java/dolla/command/AddRecurringTaskCommand.java
@@ -1,7 +1,7 @@
 package dolla.command;
 
 import dolla.Time;
-import dolla.ui.Ui;
+import dolla.Ui;
 import dolla.task.Recurring;
 import dolla.task.Task;
 import dolla.task.TaskList;
