+++ src/main/java/dolla/parser/LimitParser.java
@@ -1,5 +1,6 @@
 package dolla.parser;
 
+import dolla.command.AddLimitCommand;
 import dolla.command.Command;
 
 /**
@@ -13,8 +14,9 @@ public class LimitParser extends Parser {
 
     @Override
     public Command handleInput(String mode, String inputLine) {
-        if (commandToRun.equals("set")) {
-            //add limit command
+        if (commandToRun.equals("limits")) { //show limit list
+            //            return new ShowList(mode);
+
         } else if (commandToRun.equals("remove")) {
             //remove limit command
         } else if (commandToRun.equals("edit")) {
